Fix export port (#246)
* Add service profile for the clinic service

* Update clinic service profile

* Update clinic service profile

* Change export service port to 9301
Previously this was set to 9300 which is in linkerd's default list of opaque ports. This was causing issues with ingress traffic due to a bug in linkerd 2.12
(https://github.com/linkerd/linkerd2/issues/9888). The export service is handling http traffic so the port it listens on shouldn't be an opaque port.