make KUBECONFIG optional
The set_kubeconfig function wasn't being used anywhere, so it is removed.

The bin/tidepool script is run with the -u flag set, which causes it to abort
if any referenced variables aren't set. In the case of showing the help text,
that doesn't make sense, so a suitable default is used instead of erroring
out.