use parens to prevent exceptions with newer helm binaries
Newer version of the helm cli (e.g. v3.19.2) would error on these
expressions if the Values.deployment wasn't set.

Previous versions of the helm cli (e.g. 3.12.3) did *not* exhibit this
behavior.

These changes are tested with both 3.19.2 and 3.12.3 and seem to work
fine for both.

Thanks to Clint for point this out.